requirements.txt
Pillow
piexif
filetype
watchdog